mortAAR: Analysis of Archaeological Mortality Data

A collection of functions for the analysis of archaeological mortality data (on the topic see e.g. Chamberlain 2006 https://books.google.de/books?id=nG5FoO_becAC&lpg=PA27&ots=LG0b_xrx6O&dq=life%20table%20archaeology&pg=PA27#v=onepage&q&f=false). It takes demographic data in different formats and displays the result in a standard life table as well as plots the relevant indices (percentage of deaths, survivorship, probability of death, life expectancy, percentage of population).

See also

Author

Maintainer: Nils Mueller-Scheessel nils.mueller-scheessel@ufg.uni-kiel.de [copyright holder]

Authors:

Martin Hinz martin.hinz@ufg.uni-kiel.de

Clemens Schmid clemens@nevrome.de

Christoph Rinne crinne@ufg.uni-kiel.de

Daniel Knitter knitter@geographie.uni-kiel.de

Wolfgang Hamer hamer@geographie.uni-kiel.de

Dirk Seidensticker dirk.seidensticker@uni-tuebingen.de

Franziska Faupel ffaupel@ufg.uni-kiel.de

Carolin Tietze ctietze1991@gmail.com

Nicole Grunert nicolegrunert@gmail.com