Plots the survivorship lx for a single life table.

Plots the survivorship lx for a single life table.

Usage

mortaar_plot_lx(x, lty = 1, ...)

Arguments

x

an object of the class mortaar_life_table.

lty

line type, default to 1.

...

further arguments passed to the print function.