nitra: Life table example

The Early Neolithic cemetery of Nitra is one of the best published data sets of the Linearbandkeramik (LBK). The individuals were analysed by J. Jelínek (published in Pavúk 1972) and many decades later also bei L. Fibiger (Bickle/Whittle 2013, 139ff. Tab. 4.15) and in more detail by Z. Tvrdý 2016. We follow here the determinations by L. Fibiger.

Format

A data frame with 4 variables and 75 observations

no: individual

sex: sex determination if available

age_start: start of the ageing interval

age_end: end of the ageing interval (inclusive)

References

pavuk_1972mortAAR

LBK_lifewaysmortAAR

tvrdy_2016mortAAR

See also